cleaned up a ton

+++ TimeGoalie/app/src/main/java/us/mindbuilders/petemit/timegoalie/data/TimeGoalieDbHelper.java
@@ -4,6 +4,7 @@ import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 
+import us.mindbuilders.petemit.timegoalie.BuildConfig;
 import us.mindbuilders.petemit.timegoalie.R;
 
 /**
@@ -148,42 +149,77 @@ public class TimeGoalieDbHelper extends SQLiteOpenHelper {
 
 
 
-        //DUMMY DATA
-        // TODO: 10/6/2017 remove when done
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (1, '2017-10-05',3000,0)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-10-05',3000,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-10-04',3000,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-10-03',0,0)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-30',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-29',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-28',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-27',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-26',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-25',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-24',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-23',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-22',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-21',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-20',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-19',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-18',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-17',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-16',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-15',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-14',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-13',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-12',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-11',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-10',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-09',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-08',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-07',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-06',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-05',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-04',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-03',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-02',501,1)");
-        db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded) VALUES (4, '2017-09-01',501,1)");
+        if (BuildConfig.DEBUG) {
+            //DUMMY DATA
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (1, '2017-10-05',3000,0)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-10-05',3000,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-10-04',3000,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-10-03',0,0)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-30',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-29',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-28',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-27',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-26',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-25',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-24',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-23',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-22',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-21',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-20',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-19',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-18',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-17',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-16',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-15',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-14',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-13',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-12',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-11',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-10',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-09',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-08',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-07',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-06',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-05',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-04',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-03',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-02',501,1)");
+            db.execSQL("insert into goalentries (goal_id,timestamp,seconds_elapsed, succeeded)" +
+                    " VALUES (4, '2017-09-01',501,1)");
+        }
 
 
 
@@ -192,14 +228,16 @@ public class TimeGoalieDbHelper extends SQLiteOpenHelper {
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int i, int i1) {
-
-        //todo implement upgrade logic
-        db.execSQL("DROP TABLE IF EXISTS " + TimeGoalieContract.Goals.GOALS_TABLE_NAME);
-        db.execSQL("DROP TABLE IF EXISTS " + TimeGoalieContract.Days.DAYS_TABLE_NAME);
-        db.execSQL("DROP TABLE IF EXISTS " + TimeGoalieContract.GoalTypes.GOALTYPES_TABLE_NAME );
-        db.execSQL("DROP TABLE IF EXISTS " + TimeGoalieContract.GoalsDays.GOALS_DAYS_TABLE_NAME);
-        db.execSQL("DROP TABLE IF EXISTS " + TimeGoalieContract.GoalsDatesAccomplished.GOALS_DATES_ACCOMPLISHED_TABLE_NAME);
-        onCreate(db);
+        if (BuildConfig.DEBUG) {
+            //todo implement upgrade logic
+            db.execSQL("DROP TABLE IF EXISTS " + TimeGoalieContract.Goals.GOALS_TABLE_NAME);
+            db.execSQL("DROP TABLE IF EXISTS " + TimeGoalieContract.Days.DAYS_TABLE_NAME);
+            db.execSQL("DROP TABLE IF EXISTS " + TimeGoalieContract.GoalTypes.GOALTYPES_TABLE_NAME);
+            db.execSQL("DROP TABLE IF EXISTS " + TimeGoalieContract.GoalsDays.GOALS_DAYS_TABLE_NAME);
+            db.execSQL("DROP TABLE IF EXISTS " +
+                    TimeGoalieContract.GoalsDatesAccomplished.GOALS_DATES_ACCOMPLISHED_TABLE_NAME);
+            onCreate(db);
+        }
 
     }
